refactor: change data directory name to configs
