+++ src/configs/speakers.ts
